feat(menu): 메뉴 단일 조회 service 구현

+++ backend/foodtruck/src/main/java/com/connectruck/foodtruck/menu/dto/MenuResponse.java
@@ -1,14 +1,15 @@
 package com.connectruck.foodtruck.menu.dto;
 
 import com.connectruck.foodtruck.menu.domain.Menu;
+import java.math.BigDecimal;
 
 public record MenuResponse(
         Long id,
         String name,
-        int price
+        BigDecimal price
 ) {
 
     public static MenuResponse of(final Menu menu) {
-        return new MenuResponse(menu.getId(), menu.getName(), menu.getPrice().intValue());
+        return new MenuResponse(menu.getId(), menu.getName(), menu.getPrice());
     }
 }
